Remove logic from reusable components to single ReusableLogic and wrapped app by it
