+++ app/containers/Communities/actions.js
@@ -8,6 +8,7 @@ import {
   GET_SUGGESTED_COMMUNITIES,
   GET_SUGGESTED_COMMUNITIES_SUCCESS,
   GET_SUGGESTED_COMMUNITIES_ERROR,
+  CLEAR_SUGGESTED_COMMUNITIES,
 } from './constants';
 
 export function getSuggestedCommunities() {
@@ -27,5 +28,11 @@ export function getSuggestedCommunitiesErr(getSuggestedCommunitiesError) {
   return {
     type: GET_SUGGESTED_COMMUNITIES_ERROR,
     getSuggestedCommunitiesError,
+  };
+}
+
+export function clearSuggestedCommunities() {
+  return {
+    type: CLEAR_SUGGESTED_COMMUNITIES,
   };
 }
