+++ app/containers/Communities/constants.js
@@ -13,3 +13,6 @@ export const GET_SUGGESTED_COMMUNITIES_SUCCESS =
   'app/Communities/GET_SUGGESTED_COMMUNITIES_SUCCESS';
 export const GET_SUGGESTED_COMMUNITIES_ERROR =
   'app/Communities/GET_SUGGESTED_COMMUNITIES_ERROR';
+
+export const CLEAR_SUGGESTED_COMMUNITIES =
+  'app/Communities/CLEAR_SUGGESTED_COMMUNITIES';
