+++ app/containers/Communities/reducer.js
@@ -11,6 +11,7 @@ import {
   GET_SUGGESTED_COMMUNITIES,
   GET_SUGGESTED_COMMUNITIES_SUCCESS,
   GET_SUGGESTED_COMMUNITIES_ERROR,
+  CLEAR_SUGGESTED_COMMUNITIES,
 } from './constants';
 
 export const initialState = fromJS({
@@ -42,6 +43,12 @@ function communitiesReducer(state = initialState, action) {
       return state
         .set('getSuggestedCommunitiesLoading', false)
         .set('getSuggestedCommunitiesError', getSuggestedCommunitiesError);
+
+    case CLEAR_SUGGESTED_COMMUNITIES:
+      return state.set(
+        'suggestedCommunities',
+        initialState.get('suggestedCommunities'),
+      );
 
     case LOCATION_CHANGE:
       return initialState;
