+++ app/containers/FollowCommunityButton/constants.js
@@ -1,7 +1,2 @@
-export const FOLLOW_HANDLER = 'app/FOLLOW_BUTTON/FOLLOW_HANDLER';
-export const FOLLOW_HANDLER_SUCCESS =
-  'app/FOLLOW_BUTTON/FOLLOW_HANDLER_SUCCESS';
-export const FOLLOW_HANDLER_ERROR = 'app/FOLLOW_BUTTON/FOLLOW_HANDLER_ERROR';
-
 export const FOLLOW_BUTTON = 'follow';
 export const UNFOLLOW_BUTTON = 'unfollow';
