+++ app/containers/FollowCommunityButton/index.js
@@ -8,16 +8,10 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { createStructuredSelector } from 'reselect';
-import { compose } from 'redux';
-
-import injectSaga from 'utils/injectSaga';
-import injectReducer from 'utils/injectReducer';
 
 import { makeSelectFollowedCommunities } from 'containers/AccountProvider/selectors';
 
-import { followHandler } from './actions';
-import reducer from './reducer';
-import saga from './saga';
+import { followHandler } from 'containers/ReusableLogic/actions';
 
 /* eslint-disable react/prefer-stateless-function */
 export class FollowCommunityButton extends React.PureComponent {
@@ -60,16 +54,7 @@ function mapDispatchToProps(dispatch) /* istanbul ignore next */ {
   };
 }
 
-const withConnect = connect(
+export default connect(
   mapStateToProps,
   mapDispatchToProps,
-);
-
-const withReducer = injectReducer({ key: 'followCommunityButton', reducer });
-const withSaga = injectSaga({ key: 'followCommunityButton', saga });
-
-export default compose(
-  withReducer,
-  withSaga,
-  withConnect,
 )(FollowCommunityButton);
