--- app/containers/FollowCommunityButton/tests/actions.test.js
@@ -1,45 +1,0 @@
-import {
-  followHandler,
-  followHandlerSuccess,
-  followHandlerErr,
-} from '../actions';
-
-import {
-  FOLLOW_HANDLER,
-  FOLLOW_HANDLER_SUCCESS,
-  FOLLOW_HANDLER_ERROR,
-} from '../constants';
-
-describe('followHandler actions', () => {
-  it('FOLLOW_HANDLER', () => {
-    const communityIdFilter = 'communityIdFilter';
-    const isFollowed = 'isFollowed';
-
-    const expected = {
-      type: FOLLOW_HANDLER,
-      communityIdFilter,
-      isFollowed,
-    };
-
-    expect(followHandler(communityIdFilter, isFollowed)).toEqual(expected);
-  });
-
-  it('FOLLOW_HANDLER_SUCCESS', () => {
-    const expected = {
-      type: FOLLOW_HANDLER_SUCCESS,
-    };
-
-    expect(followHandlerSuccess()).toEqual(expected);
-  });
-
-  it('FOLLOW_HANDLER_ERROR', () => {
-    const followHandlerError = 'followHandlerError';
-
-    const expected = {
-      type: FOLLOW_HANDLER_ERROR,
-      followHandlerError,
-    };
-
-    expect(followHandlerErr(followHandlerError)).toEqual(expected);
-  });
-});
