+++ app/containers/ReusableLogic/actions.js
@@ -1,10 +1,7 @@
-/**
- *
- * VoteForNewCommunityButton
- *
- */
-
 import {
+  FOLLOW_HANDLER,
+  FOLLOW_HANDLER_SUCCESS,
+  FOLLOW_HANDLER_ERROR,
   UPVOTE,
   UPVOTE_SUCCESS,
   UPVOTE_ERROR,
@@ -12,6 +9,27 @@ import {
   DOWNVOTE_SUCCESS,
   DOWNVOTE_ERROR,
 } from './constants';
+
+export function followHandler(communityIdFilter, isFollowed) {
+  return {
+    type: FOLLOW_HANDLER,
+    communityIdFilter,
+    isFollowed,
+  };
+}
+
+export function followHandlerSuccess() {
+  return {
+    type: FOLLOW_HANDLER_SUCCESS,
+  };
+}
+
+export function followHandlerErr(followHandlerError) {
+  return {
+    type: FOLLOW_HANDLER_ERROR,
+    followHandlerError,
+  };
+}
 
 /*
  *
