+++ app/containers/ReusableLogic/constants.js
@@ -1,0 +1,15 @@
+export const FOLLOW_HANDLER = 'app/FOLLOW_BUTTON/FOLLOW_HANDLER';
+export const FOLLOW_HANDLER_SUCCESS =
+  'app/FOLLOW_BUTTON/FOLLOW_HANDLER_SUCCESS';
+export const FOLLOW_HANDLER_ERROR = 'app/FOLLOW_BUTTON/FOLLOW_HANDLER_ERROR';
+
+export const UPVOTE = 'app/SuggestedCommunities/UPVOTE';
+export const UPVOTE_SUCCESS = 'app/SuggestedCommunities/UPVOTE_SUCCESS';
+export const UPVOTE_ERROR = 'app/SuggestedCommunities/UPVOTE_ERROR';
+
+export const DOWNVOTE = 'app/SuggestedCommunities/DOWNVOTE';
+export const DOWNVOTE_SUCCESS = 'app/SuggestedCommunities/DOWNVOTE_SUCCESS';
+export const DOWNVOTE_ERROR = 'app/SuggestedCommunities/DOWNVOTE_ERROR';
+
+export const UPVOTE_BUTTON_ID = 'app_SuggestedCommunities_UPVOTE_BUTTON_ID';
+export const DOWNVOTE_BUTTON_ID = 'app_SuggestedCommunities_DOWNVOTE_BUTTON_ID';
