+++ app/containers/ReusableLogic/reducer.js
@@ -1,11 +1,14 @@
-/**
+/*
  *
- * VoteForNewCommunityButton
+ * ReusableLogic reducer
  *
  */
 
 import { fromJS } from 'immutable';
 import {
+  FOLLOW_HANDLER,
+  FOLLOW_HANDLER_SUCCESS,
+  FOLLOW_HANDLER_ERROR,
   UPVOTE,
   UPVOTE_SUCCESS,
   UPVOTE_ERROR,
@@ -15,16 +18,27 @@ import {
 } from './constants';
 
 export const initialState = fromJS({
+  followHandlerLoading: false,
+  followHandlerError: null,
   upVoteLoading: false,
   upVoteError: null,
   downVoteLoading: false,
   downVoteError: null,
 });
 
-function voteForNewCommunityButtonReducer(state = initialState, action) {
-  const { type, downVoteError, upVoteError } = action;
+function reusableLogicReducer(state = initialState, action) {
+  const { type, followHandlerError, downVoteError, upVoteError } = action;
 
   switch (type) {
+    case FOLLOW_HANDLER:
+      return state.set('followHandlerLoading', true);
+    case FOLLOW_HANDLER_SUCCESS:
+      return state.set('followHandlerLoading', false);
+    case FOLLOW_HANDLER_ERROR:
+      return state
+        .set('followHandlerLoading', false)
+        .set('followHandlerError', followHandlerError);
+
     case UPVOTE:
       return state.set('upVoteLoading', true);
     case UPVOTE_SUCCESS:
@@ -46,4 +60,4 @@ function voteForNewCommunityButtonReducer(state = initialState, action) {
   }
 }
 
-export default voteForNewCommunityButtonReducer;
+export default reusableLogicReducer;
