+++ app/containers/VoteForNewCommunityButton/constants.js
@@ -1,13 +1,2 @@
-export const UPVOTE = 'app/SuggestedCommunities/UPVOTE';
-export const UPVOTE_SUCCESS = 'app/SuggestedCommunities/UPVOTE_SUCCESS';
-export const UPVOTE_ERROR = 'app/SuggestedCommunities/UPVOTE_ERROR';
-
-export const DOWNVOTE = 'app/SuggestedCommunities/DOWNVOTE';
-export const DOWNVOTE_SUCCESS = 'app/SuggestedCommunities/DOWNVOTE_SUCCESS';
-export const DOWNVOTE_ERROR = 'app/SuggestedCommunities/DOWNVOTE_ERROR';
-
-export const UPVOTE_BUTTON_ID = 'app_SuggestedCommunities_UPVOTE_BUTTON_ID';
-export const DOWNVOTE_BUTTON_ID = 'app_SuggestedCommunities_DOWNVOTE_BUTTON_ID';
-
 export const UPVOTE_METHOD = 'upvotes';
 export const DOWNVOTE_METHOD = 'downvotes';
